+++ mazikeen/GeneratorUtils.py
@@ -1,16 +1,7 @@
 from mazikeen.GeneratorException import GeneratorException
 
 def getYamlInt(data, line, field):
-    if (not isinstance(data, int)):
-        if (isinstance(data, dict)):
-            raise GeneratorException(f"field '{field}' expects an integer at line {data['__line__']}")
-        raise GeneratorException(f"field '{field}' expects an integer at line {line}")
-
-    return data
-
-def getYamlIntOrNone(data, line, field):
-    if isinstance(data, str):
-        if data.lower() == "none": return None
+    if data == None: return None
     if (not isinstance(data, int)):
         if (isinstance(data, dict)):
             raise GeneratorException(f"field '{field}' expects an integer at line {data['__line__']}")