+++ changelog.rst
@@ -6,6 +6,8 @@ For a complete changelog, see:
 -----
 * Diff block attribute `ignoreLine` changed to `ignore`. Being able to ingnore just a portion of the line. The script file did need to adapt.
 * Run block attribute 'exitcode' no longer accepts value `None`. It is now written `null` as this is the standart
+* Run block now suppors attribute `shell`. It allwos to run a command in a selected shell. Supported shells are: `cmd`, `sh` and `powershell`
+
 1.1.3
 -----
 * mazikeen can now work also with script.yaml version 1.0.0